requirements.txt
cffi==1.13.2
cryptography==2.8
Django==2.2.4
django-rest-knox==4.1.0
djangorestframework==3.11.0
django-webpack-loader==0.7.0
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
# news scraper
appdirs==1.4.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
dateparser==0.7.6
fake-useragent==0.1.11
feedparser==5.2.1
idna==3.3
importlib-metadata==4.11.3
lxml==4.8.0
parse==1.19.0
pyee==8.2.2
pygooglenews==0.1.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
regex==2022.4.24
requests==2.27.1
requests-html==0.10.0
soupsieve==2.3.2.post1
tqdm==4.64.0
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
zipp==3.8.0
